Issues hew Issues
Took 1 hour 23 minutes

+++ lib/pages/price/hop_on_go_price_update.dart
@@ -171,10 +171,10 @@ class _HopOnGoPriceUpdateState extends State<HopOnGoPriceUpdate> {
                           if (formKey.currentState!.validate()) {
                             log("Updating");
                             await ffstore.collection("HopOnPrices").doc(widget.id).update({
-                              'pricePerMiles': int.parse(kiloMeterController.text),
+                              'pricePerMiles': double.parse(kiloMeterController.text),
                               'bookingFee':int.parse( bookingFeeController.text),
                               'serviceFee': double.parse(serviceController.text),
-                              'timeMultiplier': int.parse(timeController.text),
+                              'timeMultiplier': double.parse(timeController.text),
                             });
                             Get.snackbar(
                               "Success",
