+++ lib/pages/price/hop_on_x_price_update.dart
@@ -164,13 +164,11 @@ class _HopOnXPriceUpdateState extends State<HopOnXPriceUpdate> {
                       InkWell(
                         onTap: () async {
                           if (formKey.currentState!.validate()) {
-
-
                             await ffstore.collection("HopOnPrices").doc(widget.id).update({
-                              'pricePerMiles': int.parse(kiloMeterController.text),
-                              'bookingFee':int.parse( bookingFeeController.text),
+                              'pricePerMiles': double.parse(kiloMeterController.text),
+                              'bookingFee': int.parse(bookingFeeController.text),
                               'serviceFee': double.parse(serviceController.text),
-                              'timeMultiplier': int.parse(timeController.text),
+                              'timeMultiplier': double.parse(timeController.text),
                             });
                             Get.snackbar(
                               "Success",
